Made time of failure int. DIKERNEL-51

+++ DiKErnel/Src/DiKErnel.Core/Public/Calculator.cpp
@@ -139,14 +139,14 @@ namespace DiKErnel::Core
         for (const auto& timeDependentOutputItemsPerLocation : timeDependentOutputItems)
         {
             vector<double> damages;
-            unique_ptr<double> timeOfFailure;
+            unique_ptr<int> timeOfFailure;
             for (const auto& timeDependentOutput : timeDependentOutputItemsPerLocation)
             {
                 damages.push_back(timeDependentOutput->GetDamage());
 
                 if(timeDependentOutput->GetTimeOfFailure() != nullptr)
                 {
-                    timeOfFailure = make_unique<double>(*timeDependentOutput->GetTimeOfFailure());
+                    timeOfFailure = make_unique<int>(*timeDependentOutput->GetTimeOfFailure());
                 }
             }
